[LB-442] feat : 회원서버 트레이싱 마이크로미터 설정
