+++ .github/workflows/githubCD.yaml
@@ -1,7 +1,9 @@
-name: Build & Test
+name: Build & Delivery
 
 on:
-  push
+  push:
+    branches:
+      - 'main'
 
 env:
   REGISTRY: ghcr.io
@@ -54,6 +56,11 @@ jobs:
         run: chmod +x ./gradlew
         shell: bash
 
+      - name: Create application-dev.yaml
+        run: |
+          mkdir -p src/main/resources
+          echo "${{ vars.APPLICATION_DEV }}" > src/main/resources/application-dev.yaml
+
       - name: Setup Gradle
         uses: gradle/gradle-build-action@v2
 
@@ -62,3 +69,30 @@ jobs:
         run: ./gradlew clean build
         shell: bash
 
+      - name: Extract App Name & Version
+        run: |
+          echo "APP_NAME=$(./gradlew -q getAppName)" >> ${GITHUB_ENV};
+          echo "APP_VERSION=$(./gradlew -q getAppVersion)" >> ${GITHUB_ENV};
+
+      - name: Set Docker Image version
+        run: |
+          echo "DOCKER_IMAGE=${{ env.REGISTRY }}/${{ env.ORGANIZATION }}/${{ env.APP_NAME }}:${{ env.APP_VERSION }}" >> ${GITHUB_ENV}
+
+      - name: Login to GitHub Container Registry
+        uses: docker/login-action@v3
+        with:
+          registry: ${{ env.REGISTRY }}
+          username: ${{ env.REGISTRY_USERNAME }}
+          password: ${{ secrets.TOKEN_GITHUB }}
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Build & Push Docker Image (multi-platform)
+        uses: docker/build-push-action@v5
+        with:
+          context: .
+          file: docker/Dockerfile
+          platforms: linux/amd64,linux/arm64
+          push: true
+          tags: ${{ env.DOCKER_IMAGE }}
