+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 group = 'com.lgcms'
-version = '0.6'
+version = '0.7'
 
 java {
 	toolchain {
@@ -37,6 +37,10 @@ dependencies {
 
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 	runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
+
+	implementation 'io.micrometer:micrometer-tracing-bridge-otel' // 트레이싱
+	implementation 'io.opentelemetry:opentelemetry-exporter-otlp' // otel exporter
+	implementation 'net.ttddyy.observation:datasource-micrometer-spring-boot:1.1.2' // JDBC 트레이싱
 }
 
 tasks.named('test') {
