+++ src/main/java/com/lgcms/member/config/RedisConfig.java
@@ -1,10 +1,15 @@
 package com.lgcms.member.config;
 
+import io.lettuce.core.resource.ClientResources;
+import io.lettuce.core.tracing.MicrometerTracing;
+import io.micrometer.observation.ObservationRegistry;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+import org.springframework.data.redis.connection.lettuce.LettuceClientConfiguration;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
@@ -20,10 +25,12 @@ public class RedisConfig {
     private int database;
 
     @Bean
-    public RedisConnectionFactory redisConnectionFactory() {
-        LettuceConnectionFactory connectionFactory = new LettuceConnectionFactory(host, port);
-        connectionFactory.setDatabase(database);
-        return connectionFactory;
+    public RedisConnectionFactory redisConnectionFactory(ClientResources clientResources) {
+        LettuceClientConfiguration clientConfig = LettuceClientConfiguration.builder()
+            .clientResources(clientResources).build();
+        RedisStandaloneConfiguration redisConfig = new RedisStandaloneConfiguration(host, port);
+        redisConfig.setDatabase(database);
+        return new LettuceConnectionFactory(redisConfig, clientConfig);
     }
 
     @Bean(name = "stringTemplate")
@@ -44,5 +51,12 @@ public class RedisConfig {
         redisTemplate.setKeySerializer(new StringRedisSerializer());
         redisTemplate.setValueSerializer(new GenericJackson2JsonRedisSerializer());
         return redisTemplate;
+    }
+
+    @Bean
+    public ClientResources clientResources(ObservationRegistry observationRegistry) {
+        return ClientResources.builder()
+            .tracing(new MicrometerTracing(observationRegistry, "redis-cache"))
+            .build();
     }
 }
