+++ src/main/resources/application-local.yaml
@@ -22,26 +22,36 @@ spring:
 
   kafka:
     bootstrap-servers: localhost:38123
-
+    template:
+      observation-enabled: true
     listener:
-      ack-mode: manual_immediate
+      observation-enabled: true
+      ack-mode: manual
 
     consumer:
       group-id: ${spring.application.name}
-      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
+      key-deserializer: org.springframework.kafka.support.serializer.ErrorHandlingDeserializer
+      value-deserializer: org.springframework.kafka.support.serializer.ErrorHandlingDeserializer
       enable-auto-commit: false
-      auto-offset-reset: latest
+      auto-offset-reset: earliest
       max-poll-records: 10
       properties:
+        spring.deserializer.key.delegate.class: org.apache.kafka.common.serialization.StringDeserializer
+        spring.deserializer.value.delegate.class: org.springframework.kafka.support.serializer.JsonDeserializer
         spring.json.trusted.packages: "*"
         spring.json.use.type.headers: false
+        spring.json.value.default.type: com.lgcms.member.common.kafka.dto.KafkaEvent
 
     producer:
       key-serializer: org.apache.kafka.common.serialization.StringSerializer
       value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
       properties:
         spring.json.add.type.headers: false
+otel:
+  propagators:
+    - tracecontext
+    - b3
+    - baggage
 management:
   endpoints:
     web:
@@ -50,5 +60,8 @@ management:
   metrics:
     tags:
       application: ${spring.application.name}
+  tracing:
+    sampling:
+      probability: 1.0
 server:
   port: 38109
